[SPARK-46517][PS][TESTS] Reorganize `IndexingTest`: factor out `test_loc*` tests
### What changes were proposed in this pull request?
1, factor out `test_loc*` tests
2, add the missing parity tests
(will fix remaining parts in followups)

### Why are the changes needed?
1, for test parity
2, for parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44502 from zhengruifeng/ps_test_indexing_loc.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
